+++ app/routes/routes.js
@@ -1,700 +1,30 @@
-// app/routes.js
 
-//load bcrypt
-var bCrypt = require('bcrypt-nodejs');
-var crypto = require('crypto');
-var jwt = require('jsonwebtoken');
-var randomstring = require('randomstring');
 
 module.exports = (app, models) => {
 
-    // =====================================
-    // HOME PAGE (with login links) ========
-    // =====================================
-
-    var Client = models.client;
-    var Vendor = models.vendor;
-
-    var ResponseFormat = function() {
-        this.status = '',
-        this.message = '',
-        this.data = ''
-    };
-
-
-    var generateHash = function(password) {
-        return bCrypt.hashSync(password, bCrypt.genSaltSync(8), null);
-    };
-
-    var isValidPassword = function(userpass, password) {
-        return bCrypt.compareSync(password, userpass);
-    };
-
-    app.get('/', (req, res) => {
-        res.render('index.ejs'); // load the index.ejs file
-    });
-
-    // =====================================
-    // LOGIN ===============================
-    // =====================================
-    // show the login form
-    app.get('/login', (req, res) => {
-
-        // render the page and pass in any flash data if it exists
-        res.render('login.ejs', { message: req.flash('loginMessage') });
-    });
-
-    // process the login form
-    app.post('/login', (req, res, next) => {
-
-        var temp = new ResponseFormat();
-
-        var User;
-
-        if (req.body.type == 'client') {
-            console.log('client');
-
-            User = Client;
-        }
-        else if (req.body.type == 'vendor'){
-            console.log('vendor');
-        
-            User = Vendor;
-        }
-
-        User.findOne({where: { email: req.body.username }}).then(function(user) {
-
-            if (!user) {
-                temp.status = 401;
-                temp.message = 'Authentication Failed. User not found!';
-                temp.data = null;
-
-                res.status(temp.status)
-                    .json(temp);
-            }
-
-            else if (user) {
-                
-                if (!isValidPassword(user.password, req.body.password)) { 
-                    temp.status = 401;
-                    temp.message = 'Authentication Failed. Password Incorrect!';
-                    temp.data = null; 
-
-                     res.status(temp.status)
-                        .json(temp);
-                }
-                else {
-                
-                        var Verify = models.verification;
-
-                        Verify.findOne({where: {id : user.id}}).then(function(verify) {
-                        
-                                console.log(verify);
-                                if (!verify.accountVerified) {
-                                    temp.status = 403;
-                                    temp.message = 'You are not verified, please verify your account.';
-                                    temp.data = verify; 
-
-                                    return res.status(temp.status).json(temp);    
-                                }
-                                temp.status = 200;
-                                temp.message = 'User logged in Successfully';
-                                temp.data = {"token": jwt.sign({ email: user.email, id: user.id, type: req.body.type }, 'ClientVendor')};
-
-                                return res.status(temp.status).json(temp);
-                        });
-                }
-            }
-        })
-    });
-
-
-    // =====================================
-    // SIGNUP ==============================
-    // =====================================
-    // show the signup form
-    app.get('/signup', (req, res) => {
-        // render the page and pass in any flash data if it exists
-        res.render('signup.ejs', { message: req.flash('signupMessage') });
-    });
-
-    // process the login form
-    app.post('/signup', (req, res, next) => {
-
-        var temp = new ResponseFormat();
-
-        var User;
-
-        if (req.body.type == 'client') {
-            console.log('client');
-
-            User = Client;
-        }
-        else if (req.body.type == 'vendor'){
-            console.log('vendor');
-        
-            User = Vendor;
-        }
-
-        User.findOne({ where: { email: req.body.username }}).then(function(user) {
-
-            if (user)
-            {
-                temp.status = 200;
-                temp.message = 'User already exists.';
-                temp.data = null;
-
-                res.status(temp.status)
-                    .json(temp);
-            } 
-            else {
-
-                var userPassword = generateHash(req.body.password);
-                const random_id = crypto.randomBytes(16).toString('hex');
-                console.log(random_id);
-
-                var data = {
-                    id: random_id, 
-                    email: req.body.username,
-                    password: userPassword,
-                    name: req.body.name,
-                    phoneNumber: req.body.phoneNumber
-                };
-
-                User.create(data).then(function(newUser) {
-                    
-                    if (!newUser) {
-                        temp.status = 200;
-                        temp.message = 'Unable to create the user.'
-                        temp.data = null;
-                    }
-
-                    if (newUser) {
-                        temp.status = 200;
-                        temp.message = 'User created Successfully.'
-                        temp.data = newUser;
-
-                        var permalink_local = req.body.username.toLowerCase().replace(' ', '').replace(/[^\w\s]/gi, '').trim();
-
-                        var token = randomstring.generate({
-                            length: 64
-                        });
-
-                        var link_data = {
-                            id : newUser.id,
-                            verify_token : token,
-                            permalink : permalink_local,
-                            accountVerified : false
-                        }
-                        console.log(permalink_local +token);
-
-                        var link = "http://localhost:8080/verification/" + permalink_local + "/" + token + "/" + newUser.id;
-
-                        var Verification = models.verification;
-                        temp.status = 200;
-
-                        Verification.create(link_data).then(function(client) {
-
-                            if (!client) {
-                                console.log("error");
-                                temp.message = 'error with verification process';
-                                temp.data = null;
-
-                                return res.status(temp.status).json(temp);
-                            }
-
-                            //console.log("added to verification");
-                            temp.message = 'Successful Signup and link generated';
-                            temp.data = newUser;
-                            console.log(temp.message);
-                        });
-
-                        var mailOptions = {
-                            to: newUser.email,
-                            subject: 'Verification link',
-                            user: {
-                                login_link: link
-                            },
-                            attachments: [{   // file on disk as an attachment
-                                filePath: "../../data/TnC.pdf" // stream this file
-                            }]
-                        }
-                        if (req.body.type == 'client')
-                        {
-                            app.mailer.send('email', mailOptions, function (err, message) {
-                                if (err) {
-                                    console.log(err);
-                                    return;
-                                }
-                                console.log("mail sent");
-                                temp.data = {"id": newUser.id, "email":newUser.email, "type": req.body.type};
-                                return res.status(temp.status).json(temp);  
-                            });
-                        }
-                        else
-                        {
-                            temp.message = 'Signup successfull, contact the admin for verification';
-                            temp.data = {"id": newUser.id, "email":newUser.email, "type": req.body.type};
-                            return res.status(temp.status).json(temp);
-                        }
-                        
-                    }
-
-                    
-                });
-            }
-        }); 
-    });
-
-
-    app.get('/verification/:link/:token/:id', (req, res) => {
-
-        var temp = new ResponseFormat();
-
-        var Verification = models.verification;
-
-        Verification.update({accountVerified : true}, {where :{id : req.params.id, $and:[
-            {permalink : req.params.link}, 
-            {verify_token : req.params.token},
-            {accountVerified : false}
-            ]}}).then(function(client) {
-
-            if(!client) {
-                temp.status = 500;
-                temp.message = 'verification unsuccessfull';
-                temp.data = null;
-
-                return res.status(temp.status)
-                    .json(temp);
-            }
-
-            temp.status = 200;
-            temp.message = 'The email for client ' + req.params.id + ' is verified.';
-            temp.data = client;
-
-            return res.status(temp.status).json(temp);
-        });
-
-    });
-    
-    // =====================================
-    // PROFILE SECTION =========================
-    // =====================================
-    // we will want this protected so you have to be logged in to visit
-    // we will use route middleware to verify this (the isLoggedIn function)
-    app.get('/profile', isLoggedIn, ({ user }, res) => {
-        // res.render('profile.ejs', {
-        //     user, // get the user out of session and pass to template
-        // });
-        // res.send(user);
-        console.log("This is the data to be displayed");
-        console.log(user);
-    });
-
-    //Route to get all the Main Categories
-    app.get('/main_categories', isLoggedIn, (req, res) => {
-
-        var temp = new ResponseFormat();
-        var MainCategory = models.main_category;
-
-        MainCategory.findAll().then(function(mainCategories) {
-                
-            temp.status = 200;
-
-            if (!mainCategories.length) {
-                temp.message = 'No Main Categories Found';
-                temp.data = [];    
-            }
-            else { 
-                temp.message = 'All the Main Categories';
-                temp.data = mainCategories;   
-            }
-
-            res.status(temp.status)
-                .json(temp);
-        })
-    });
-
-
-    //Route to create a main category...this will only be used by developers and Admin
-    app.post('/main_categories', function(req, res) {
-
-        var temp = new ResponseFormat();
-
-        var MainCategory = models.main_category;
-
-            var data = {
-                name: req.body.name
-            };
-
-            MainCategory.create(data).then(function(mainCategory) {
-
-                if (mainCategory) {
-                    temp.status = 201;
-                    temp.message = 'Successfully created the categorY';
-                    temp.data = mainCategory;
-                }
-                else {
-                    temp.status = 409;
-                    temp.message = 'Unable to create the category';
-                    temp.data = [];
-                }
-
-                res.status(temp.status)
-                    .json(temp);
-            });
-    });
-    //Route to get all the Sub Categories associated with a Main Category
-    app.get('/sub_categories/:id', isLoggedIn, function(req, res) {
-
-        var temp = new ResponseFormat();
-        var SubCategory = models.sub_category;
-
-        SubCategory.findAll({where: {mainCategoryId : req.params.id}}).then(function(subCategories) {
-
-            temp.status = 200;
-            
-            if (!subCategories.length) {
-                temp.message = 'No Sub Categories Found';
-                temp.data = [];
-            }
-            else {
-                temp.message = 'Sub categories corresponding to Main Category #' + req.params.id;
-                temp.data = subCategories;   
-            }
-
-            res.status(temp.status)
-                .json(temp);
-        })
-    });
-
-    //Route to create sub categories for a main category...for Developers and Admin usage only
-    app.post('/sub_categories', function(req, res) {
-
-        var temp = new ResponseFormat();
-        var SubCategory = models.sub_category;
-
-        var data = {
-                name: req.body.name,
-                mainCategoryId: req.body.mainCategoryId
-            };
-
-        SubCategory.create(data).then(function(subCategory) {
-
-            if (subCategory) {
-                    temp.status = 201;
-                    temp.message = 'Successfully created the sub category';
-                    temp.data = subCategory;
-                }
-                else {
-                    temp.status = 409;
-                    temp.message = 'Unable to create the sub category';
-                    temp.data = [];
-                }
-
-                res.status(temp.status)
-                    .json(temp);
-        })
-    });
-
-    app.post('/tender', isLoggedIn, function(req, res) {
-
-        var temp = new ResponseFormat();
-
-        var Tender = models.tender;
-
-        if (req.user.id && req.body.duration && req.body.quantity && req.body.subCategoryId) {
-
-            console.log("Client:  " + req.user.id + " is posting a Tender. " + "Duration: " + req.body.duration + " Quantity: " + req.body.quantity + " SubCategory: " + req.body.subCategoryId);
-
-            var tenderData = {
-                tenderEnds: req.body.duration,
-                quantity: req.body.quantity,
-                clientId: req.user.id,
-                subCategoryId: req.body.subCategoryId
-            };
-
-            Tender.create(tenderData).then(function(newTender) {
-
-                if (newTender) {
-                    temp.status = 201;
-                    temp.message = 'Successfully created the Tender';
-                    temp.data = newTender;
-                }
-                else {
-                    temp.status = 409;
-                    temp.message = 'Unable to create the Tender';
-                    temp.data = [];
-                }
-
-                res.status(temp.status)
-                    .json(temp);
-            });
-        }
-        else
-        {
-            temp.status = 403;
-            temp.message = 'Client needs to be logged in to post a Tender';
-            temp.data = null;
-
-            res.status(temp.status)
-                .json(temp);
-        }
-    });
-
-    // =====================================
-    // LOGOUT ==============================
-    // =====================================
-    app.get('/logout', (req, res) => {
-        req.logout();
-        res.redirect('/');
-    });
-
-    app.get('/client_tenders', isLoggedIn, function(req, res) {
-
-        var temp = new ResponseFormat();
-
-        var Tender = models.tender;
-
-        if (req.user.id) {
-
-            Tender.findAll({where: {clientId: req.user.id}}).then(function(clientTenders) {
-
-                if (clientTenders.length) {
-                    temp.status = 200;
-                    temp.message = 'Retreived all Tenders for Client ' + req.user.id;
-                    temp.data = clientTenders;
-                }
-                else {
-                    temp.status = 200;
-                    temp.message = 'Unable to find Tenders posted by Client ' + req.user.id;
-                    temp.data = null;
-                }
-                
-                res.status(temp.status)
-                    .json(temp);
-            })
-        }
-        else
-        {
-            temp.status = 403;
-            temp.message = 'Client needs to be logged in to retrieve all the Tenders';
-            temp.data = null;
-                
-            res.status(temp.status)
-                .json(temp);
-        }
-    });
-
-    app.get('/tenders_main_category', function(req, res) {
-
-        var temp = new ResponseFormat();
-
-        var Tender = models.tender;
-        var SubCategory = models.sub_category;
-
-        var mainCategoryIds = req.query.mainCategoryId;
-
-        if (!Array.isArray(req.query.mainCategoryId)){
-            mainCategoryIds = Array.from(req.query.mainCategoryId);
-        }   
-
-        Tender.findAll({
-            include: [{
-                model: models.sub_category,
-                where: {mainCategoryId: {in: mainCategoryIds}}
-            }]
-        }).then(function(tenders) {
-            
-            temp.status = 200;
-                
-            if (tenders.length) {
-                temp.message = 'Tenders associated with the requested Main Categories';
-                temp.data = tenders;
-            }
-            else {
-                temp.message = 'No tenders for the requested Main Categories';
-                temp.data = null;
-            }
-
-            res.status(temp.status)
-                .json(temp);
-        });
-    });
-
-
-    app.get('/basic_details', isLoggedIn, function(req, res) {
-
-        var temp = new ResponseFormat();
-
-        var User = null;
-
-        if (req.user.type == 'client') {
-            User = models.client;
-        }
-        else if (req.user.type == 'vendor') {
-            User = models.vendor;
-        }
-
-        if (User != null) {
-
-            User.findById(req.user.id, {attributes: ['name', 'phoneNumber', 'email']}).then(function(user) {
-
-                if (user) {
-                    temp.status = 200;
-                    temp.message = 'Basic Profile for ' + req.user.email;
-                    temp.data = user;
-                }
-                else {
-                    temp.status = 200;
-                    temp.message = 'Unable to retrieve the Basic Profile for ' + req.user.email;
-                    temp.data = null;
-                }
-
-                res.status(temp.status)
-                    .json(temp);
-            });
-        }
-    });
-
-    app.put('/basic_details', isLoggedIn, function(req, res) {
-
-        var temp = new ResponseFormat();
-
-        var User = null;
-
-        if (req.user.type == 'client') {
-            User = models.client;
-        }
-        else if (req.user.type == 'vendor') {
-            User = models.vendor;
-        }
-
-        if (User != null) {
-
-            if (req.body.name && req.body.phoneNumber && req.body.email) {
-
-                User.update({name: req.body.name, phoneNumber: req.body.phoneNumber, email: req.body.email}, {where: {id: req.user.id}}).then(function(updatedUser) {
-                    if (updatedUser){
-
-                        User.findById(req.user.id, {attributes: ['name', 'phoneNumber', 'email']}).then(function(user) {
-
-                            if (user) {
-                                temp.status = 200;
-                                temp.message = 'Basic Profile updated Successfully'
-                                temp.data = user;
-                            }
-                            else {
-                                temp.status = 200;
-                                temp.message = 'Unable to find the Updated User'
-                                temp.data = null;
-                            }
-
-                            res.status(temp.status)
-                                .json(temp);
-                        });
-                    }
-                    else {
-                        temp.status = 400;
-                        temp.message = 'Something went wrong with the update'
-                        temp.data = null;
-
-                        res.status(temp.status)
-                            .json(temp);
-                    }
-                })
-            }
-            else {
-                temp.status = 422;
-                temp.message = 'Missing Parameters!';
-                temp.data = req.body;
-
-                res.status(temp.status)
-                    .json(temp);
-            }
-        }
-    });
-
-    app.put('/business_details', isLoggedIn, function(req, res) {
-
-        var temp = new ResponseFormat();
-        var BusinessDetails = models.business_details;
-
-        if (req.body.bankName && req.body.ifscCode && req.body.bankBranch && req.body.address && req.body.gstNumber && req.body.accountNumber) {
-
-            var tempBusinessDetails = {
-                id: req.user.id,
-                bankName: req.body.bankName,
-                ifscCode: req.body.ifscCode,
-                bankBranch: req.body.bankBranch,
-                address: req.body.address,
-                gstNumber: req.body.gstNumber,
-                accountNumber: req.body.accountNumber
-            };
-
-            BusinessDetails.upsert(tempBusinessDetails).then(function(created) {
-                
-                BusinessDetails.findById(req.user.id).then(function(user) {
-
-                    if (user) {
-                        temp.status = 200;
-                        temp.message = 'Business Profile updated Successfully'
-                        temp.data = user;
-                    }
-                    else {
-                        temp.status = 400;
-                        temp.message = 'Unable to find the Updated User'
-                        temp.data = null;
-                    }
-
-                    res.status(temp.status)
-                        .json(temp);
-                });
-            });
-        }
-        else {
-            temp.status = 422;
-            temp.message = 'Missing Parameters!';
-            temp.data = req.body;
-
-            res.status(temp.status)
-                .json(temp);
-        }
-
-    })
-
-
-    app.get('/business_details', isLoggedIn, function(req, res) {
-
-        var temp = new ResponseFormat();
-
-        var BusinessDetails = models.business_details;
-
-        BusinessDetails.findById(req.user.id).then(function(user) {
-
-            if (user) {
-                temp.status = 200;
-                temp.message = 'Business Profile for ' + req.user.email;
-                temp.data = user;
-            }
-            else {
-                temp.status = 200;
-                temp.message = 'Unable to retrieve the Business Profile for ' + req.user.email
-                temp.data = null;
-            }
-
-            res.status(temp.status)
-                .json(temp);
-        });
-    });
-};
-
-// route middleware to make sure
-function isLoggedIn(req, res, next) {
-
-    // if user is authenticated in the session, carry on
-    if (req.user) return next();
-
-    // if they aren't redirect them to the home page
-    res.redirect('/');
+	var common = require('./common.js')
+	var authentication = require('./authentication.js')(models, common)
+	var categories = require('./categories.js')(models, common)
+	var profile_update = require('./profile_update.js')(models, common)
+	var tenders = require('./tenders.js')(models, common)
+
+	app.post('/login', authentication.login)
+	app.post('/signup', authentication.signup)
+	app.get('/verification/:link/:token/:id', authentication.verify)
+
+	app.get('/basic_details', common.isLoggedIn, profile_update.get_basic_details())
+		.put('/basic_details', common.isLoggedIn, profile_update.update_basic_details())
+
+	app.get('/business_details', common.isLoggedIn, profile_update.get_business_details())
+		.put('/business_details', common.isLoggedIn, profile_update.update_business_details())
+
+	app.post('/tender', common.isLoggedIn, tenders.add_tender())
+	app.get('/client_tenders', common.isLoggedIn, tenders.get_client_tenders())
+	app.get('/tenders_main_category', common.isLoggedIn, tenders.get_main_category_tenders())
+
+	app.get('/main_categories', common.isLoggedIn, categories.get_main_categories())
+		.post('/main_categories', categories.add_main_category())
+
+	app.get('/sub_categories/:id', common.isLoggedIn, categories.get_sub_categories())
+	app.post('/sub_categories', categories.add_sub_category())
 }
