+++ server.js
@@ -76,7 +76,7 @@ mailer.extend(app, {
 });
 
 // routes ======================================================================
-require('./app/routes/routes.js')(app, models); // load our routes and pass in our app and fully configured passport
+require('./app/routes/routes.js')(app, models); // load our routes and pass in our app
 
 // launch ======================================================================
 app.listen(port);
